requirements.txt
google-adk
openai>=1.0.0
google-cloud-aiplatform>=1.0.0
gradio>=6.0.0
litellm
requests
